move things around CLIs
