+++ src/databricks/labs/ucx/contexts/application.py
@@ -1,7 +1,5 @@
 import abc
 import logging
-import os
-import shutil
 from datetime import timedelta
 from functools import cached_property
 
@@ -13,17 +11,8 @@ from databricks.sdk import AccountClient, WorkspaceClient, core
 from databricks.sdk.service import sql
 
 from databricks.labs.ucx.account import WorkspaceInfo
-from databricks.labs.ucx.assessment.aws import AWSResources, run_command
 from databricks.labs.ucx.assessment.azure import AzureServicePrincipalCrawler
-from databricks.labs.ucx.aws.access import AWSResourcePermissions
-from databricks.labs.ucx.aws.credentials import CredentialManager, IamRoleMigration
-from databricks.labs.ucx.azure.access import AzureResourcePermissions
-from databricks.labs.ucx.azure.credentials import (
-    ServicePrincipalMigration,
-    StorageCredentialManager,
-)
-from databricks.labs.ucx.azure.locations import ExternalLocationsMigration
-from databricks.labs.ucx.azure.resources import AzureAPIClient, AzureResources
+from databricks.labs.ucx.aws.credentials import CredentialManager
 from databricks.labs.ucx.config import WorkspaceConfig
 from databricks.labs.ucx.hive_metastore import ExternalLocations, Mounts, TablesCrawler
 from databricks.labs.ucx.hive_metastore.catalog_schema import CatalogSchema
@@ -235,56 +224,8 @@ class GlobalContext(abc.ABC):
         return AzureServicePrincipalCrawler(self.workspace_client, self.sql_backend, self.inventory_database)
 
     @cached_property
-    def azure_cli_authenticated(self):
-        if not self.is_azure:
-            raise NotImplementedError("Azure only")
-        if self.connect_config.auth_type != "azure-cli":
-            raise ValueError("In order to obtain AAD token, Please run azure cli to authenticate.")
-        return True
-
-    @cached_property
-    def azure_subscription_id(self):
-        subscription_id = self.named_parameters.get("subscription_id")
-        if not subscription_id:
-            raise ValueError("Please enter subscription id to scan storage accounts in.")
-        return subscription_id
-
-    @cached_property
-    def azure_management_client(self):
-        if not self.azure_cli_authenticated:
-            raise NotImplementedError
-        return AzureAPIClient(
-            self.workspace_client.config.arm_environment.resource_manager_endpoint,
-            self.workspace_client.config.arm_environment.service_management_endpoint,
-        )
-
-    @cached_property
-    def microsoft_graph_client(self):
-        if not self.azure_cli_authenticated:
-            raise NotImplementedError
-        return AzureAPIClient("https://graph.microsoft.com", "https://graph.microsoft.com")
-
-    @cached_property
     def external_locations(self):
         return ExternalLocations(self.workspace_client, self.sql_backend, self.inventory_database)
-
-    @cached_property
-    def azure_resources(self):
-        # TODO: move to cli_command.py
-        return AzureResources(
-            self.azure_management_client,
-            self.microsoft_graph_client,
-            [self.azure_subscription_id],
-        )
-
-    @cached_property
-    def azure_resource_permissions(self):
-        return AzureResourcePermissions(
-            self.installation,
-            self.workspace_client,
-            self.azure_resources,
-            self.external_locations,
-        )
 
     @cached_property
     def azure_acl(self):
@@ -319,80 +260,8 @@ class GlobalContext(abc.ABC):
         )
 
     @cached_property
-    def aws_cli_run_command(self):
-        # TODO: slowly move this to cli_command.py
-        # this is a convenience method for unit testing
-        if not shutil.which("aws"):
-            raise ValueError("Couldn't find AWS CLI in path. Please install the CLI from https://aws.amazon.com/cli/")
-        return run_command
-
-    @cached_property
-    def aws_profile(self):
-        # TODO: slowly move this to cli_command.py
-        aws_profile = self.named_parameters.get("aws_profile")
-        if not aws_profile:
-            aws_profile = os.getenv("AWS_DEFAULT_PROFILE")
-        if not aws_profile:
-            raise ValueError(
-                "AWS Profile is not specified. Use the environment variable [AWS_DEFAULT_PROFILE] "
-                "or use the '--aws-profile=[profile-name]' parameter."
-            )
-        return aws_profile
-
-    @cached_property
-    def aws_resources(self):
-        if not self.is_aws:
-            raise NotImplementedError("AWS only")
-        return AWSResources(self.aws_profile, self.aws_cli_run_command)
-
-    @cached_property
-    def aws_resource_permissions(self):
-        return AWSResourcePermissions(
-            self.installation,
-            self.workspace_client,
-            self.sql_backend,
-            self.aws_resources,
-            self.external_locations,
-            self.inventory_database,
-            self.named_parameters.get("aws_account_id"),
-            self.named_parameters.get("kms_key"),
-        )
-
-    @cached_property
     def iam_credential_manager(self):
         return CredentialManager(self.workspace_client)
-
-    @cached_property
-    def iam_role_migration(self):
-        return IamRoleMigration(
-            self.installation,
-            self.workspace_client,
-            self.aws_resource_permissions,
-            self.iam_credential_manager,
-        )
-
-    @cached_property
-    def azure_credential_manager(self):
-        return StorageCredentialManager(self.workspace_client)
-
-    @cached_property
-    def service_principal_migration(self):
-        return ServicePrincipalMigration(
-            self.installation,
-            self.workspace_client,
-            self.azure_resource_permissions,
-            self.azure_service_principal_crawler,
-            self.azure_credential_manager,
-        )
-
-    @cached_property
-    def azure_external_locations_migration(self):
-        return ExternalLocationsMigration(
-            self.workspace_client,
-            self.external_locations,
-            self.azure_resource_permissions,
-            self.azure_resources,
-        )
 
     @cached_property
     def table_mapping(self):
